refactor: extract table display-width helpers into src/table.ts
process-manager.ts mixed process lifecycle management with table
rendering string-width logic, making both harder to follow and test
in isolation.
